Address SonarQube issues
